+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Release Versions:
 
 - feat(controllers)!: remove robot description parameter (#186)
 - feat(controllers): add TF listener interface in BaseControllerInterface (#169)
+- feat(controllers): add TF broadcaster in BaseControllerInterface (#170)
 
 ## 5.1.0